TIGER-1256: Implement XSS sniff for Magento 1 (#10)
 - Implemented XSS sniff for M1
 - Implemented unit tests for XSS sniff for M1
 - Fixed XSS sniff for M2
 - Changed unit tests for XSS sniff for M2
